fix: score multiplier regression — restore old 1/2/3 values, unify save/restore
The magic-number refactor accidentally changed GameController.score_multiplier
from the original {'easy':1,'medium':2,'hard':3} to the milestone-difficulty
multiplier (1.0, 1.5, 2.0), reducing medium score gain by 25% and hard by 33%.

Fix: update DIFFICULTY_MULTIPLIERS in constants.py to (1.0, 2.0, 3.0) matching
the original score multiplier values (which also correctly scales milestone
thresholds). Also make restore_from_save use the same function instead of
a second hardcoded dict, and fix score_multiplier type hints int→float.

+++ airwar/game/managers/collision_controller.py
@@ -153,7 +153,7 @@ class CollisionController:
         explosive_level: int = 0,
         piercing_level: int = 0,
         player_invincible: bool = False,
-        score_multiplier: int = 1,
+        score_multiplier: float = 1.0,
         on_enemy_killed: Callable[[int], None] = None,
         on_boss_killed: Callable[[int], None] = None,
         on_boss_hit: Callable[[int], None] = None,
@@ -231,7 +231,7 @@ class CollisionController:
         self,
         player_bullets: List['Bullet'],
         enemies: List['Enemy'],
-        score_multiplier: int,
+        score_multiplier: float,
         explosive_level: int,
         piercing_level: int = 0,
     ) -> Tuple[int, int]:
